feat: add function to system service

+++ lk_flow/__main__.py
@@ -29,14 +29,28 @@ def run() -> None:  # pragma: no cover
     asyncio.get_event_loop().run_until_complete(start_server())
 
 
-def init() -> None:
-    """初始化系统"""
+def init(
+    systemd: bool = True, create_command: bool = True, work_dir: str = None
+) -> None:
+    """初始化系统
+
+    Args:
+        systemd: 增加到system service
+        create_command: 创建软连接
+        work_dir: systemd service工作目录 默认当前目录
+    """
     from lk_flow.config import conf
     from lk_flow.core import Context, mod_init
+    from lk_flow.utils import add_systemd, ln_command
 
+    generate_config(work_dir=work_dir, force=False)
     context = Context(config=conf)
     loading_plugin(context.config.mod_dir)
     mod_init(context)
+    if systemd:
+        add_systemd(work_dir)
+    if create_command:
+        ln_command()
 
 
 def entry_point() -> None:  # pragma: no cover
@@ -58,15 +72,25 @@ def entry_point() -> None:  # pragma: no cover
     fire.Fire(command_map)
 
 
-def generate_config() -> str:
-    """生成配置文件至当前目录"""
+def generate_config(work_dir: str = None, force: bool = False) -> str:
+    """生成配置文件至当前目录
+
+    Args:
+        work_dir: 生成目录
+        force: 强制重写
+    """
     import os
     import shutil
 
+    if work_dir is None:
+        work_dir = os.getcwd()
     config_file = os.path.join(os.path.dirname(__file__), "etc", "lk_flow_config.yaml")
-    target_file = os.path.join(os.getcwd(), "lk_flow_config.yaml")
-    shutil.copy(config_file, target_file)
-    return f"generate to {target_file}"
+    target_file = os.path.join(work_dir, "lk_flow_config.yaml")
+    if os.path.exists(target_file) and not force:
+        return f"already exists at {target_file}"
+    else:
+        shutil.copy(config_file, target_file)
+        return f"generate to {target_file}"
 
 
 if __name__ == "__main__":
